mcp-server/src/server.py
import asyncio
from python_mcp.server import Server
from python_mcp.models import Tool
import json
from typing import Dict, Any, List
from dotenv import load_dotenv
import os

# Load environment variables
load_dotenv()

class TodoMCPServer:
    def __init__(self):
        self.server = Server()
        self._register_tools()

    def _register_tools(self):
        """
        Register all MCP tools for todo management.
        Actual implementations will be loaded from the tools directory.
        """
        # Tools will be registered dynamically
        pass

    async def start(self, host: str = "localhost", port: int = 8080):
        """
        Start the MCP server
        """
        await self.server.start(host, port)
        print(f"MCP Server started on {host}:{port}")

    async def stop(self):
        """
        Stop the MCP server
        """
        await self.server.stop()
        print("MCP Server stopped")


# Example usage
async def main():
    server = TodoMCPServer()
    try:
        await server.start()
        # Keep the server running
        while True:
            await asyncio.sleep(1)
    except KeyboardInterrupt:
        print("Shutting down server...")
        await server.stop()


if __name__ == "__main__":
    asyncio.run(main())