mcp-server/pyproject.toml
[tool.poetry]
name = "todo-mcp-server"
version = "0.1.0"
description = "MCP Server for Todo AI Chatbot"
authors = ["Developer <developer@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
python-mcp = "^1.0.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]